chore: MemberOutfit 관련 객체 분리
